Initial commit for open source release.
Add Open Iconic icons and update project structure

This commit introduces the Open Iconic icon set, including the addition of FONT-LICENSE and ICON-LICENSE files to clarify licensing terms. The README.md file has been enhanced with project details and usage instructions. Several font files were added to support the icon set, and the site.css file was updated to include styles for the icons and other UI elements. Additionally, a favicon.png was added for branding, and the VgaUI.sln file was updated to reflect the current project structure and dependencies.